webpack-dev-server serving through API for better control and adjustments
